requirements.txt
Flask==2.3.3
gunicorn==21.0.1
python-dotenv==1.0.0
PyTumblr==0.1.2
requests-oauthlib==1.3.1
tumblrdotcom==0.1.1